End of day 6

+++ Application/Activities/Commands/DeleteActivity.cs
@@ -1,4 +1,5 @@
-﻿using AutoMapper;
+﻿using Application.Core;
+using AutoMapper;
 using MediatR;
 using Persistence;
 
@@ -6,18 +7,23 @@ namespace Application.Activities.Commands
 {
 	public class DeleteActivity
 	{
-		public class Command : IRequest
+		public class Command : IRequest<Result<Unit>>
 		{
 			public required string Id { get; set; }
 		}
 
-		public class Handler(AppDbContext _context) : IRequestHandler<Command>
+		public class Handler(AppDbContext _context) : IRequestHandler<Command, Result<Unit>>
 		{
-			public async Task Handle(Command request, CancellationToken cancellationToken)
+			public async Task<Result<Unit>> Handle(Command request, CancellationToken cancellationToken)
 			{
-				var activity = await _context.Activities.FindAsync(request.Id, cancellationToken) ?? throw new Exception("Activity not found");
+				var activity = await _context.Activities.FindAsync(request.Id, cancellationToken);
+				if (activity == null) return Result<Unit>.Failure("Activity not found", 404);
+
 				_context.Activities.Remove(activity);
-				await _context.SaveChangesAsync(cancellationToken);
+				var deleted = await _context.SaveChangesAsync(cancellationToken) > 0;
+				if (!deleted) return Result<Unit>.Failure("Failed to delete activity", 400);
+
+				return Result<Unit>.Success(Unit.Value);
 			}
 		}
 	}
